Add file input reading

+++ bbx_dsp/src/reader.rs
@@ -1,7 +1,7 @@
 use crate::sample::Sample;
 
 pub trait Reader<S: Sample>: Send + Sync {
-    fn sample_rate(&self) -> usize;
+    fn sample_rate(&self) -> f64;
     fn num_channels(&self) -> usize;
     fn num_samples(&self) -> usize;
